+++ src/Config/DriverConfig.php
@@ -22,6 +22,14 @@ use Cycle\Database\Driver\DriverInterface;
  */
 abstract class DriverConfig
 {
+    use RestoreStateTrait;
+
+    protected array $defaultOptions = [
+        'withDatetimeMicroseconds' => false,
+    ];
+
+    public array $options = [];
+
     /**
      * @param T $connection
      * @param class-string<DriverInterface> $driver
@@ -31,6 +39,9 @@ abstract class DriverConfig
      * @param bool $queryCache Enables query caching
      * @param bool $readonlySchema Disable schema modifications
      * @param bool $readonly Disable write expressions
+     * @param array{
+     *     withDatetimeMicroseconds?: bool
+     * } $options
      */
     public function __construct(
         public ConnectionConfig $connection,
@@ -40,6 +51,8 @@ abstract class DriverConfig
         public bool $queryCache = true,
         public bool $readonlySchema = false,
         public bool $readonly = false,
+        array $options = []
     ) {
+        $this->options = $options + $this->defaultOptions;
     }
 }