+++ src/Driver/MySQL/Schema/MySQLColumn.php
@@ -16,10 +16,20 @@ use Cycle\Database\Exception\DefaultValueException;
 use Cycle\Database\Injection\Fragment;
 use Cycle\Database\Injection\FragmentInterface;
 use Cycle\Database\Schema\AbstractColumn;
+use Cycle\Database\Schema\Attribute\ColumnAttribute;
 
 /**
  * Attention! You can use only one timestamp or datetime with DATETIME_NOW setting! Thought, it will
  * work on multiple fields with MySQL 5.6.6+ version.
+ *
+ * @method $this|AbstractColumn primary(int $size, bool $unsigned = false, $zerofill = false)
+ * @method $this|AbstractColumn bigPrimary(int $size, bool $unsigned = false, $zerofill = false)
+ * @method $this|AbstractColumn integer(int $size, bool $unsigned = false, $zerofill = false)
+ * @method $this|AbstractColumn tinyInteger(int $size, bool $unsigned = false, $zerofill = false)
+ * @method $this|AbstractColumn smallInteger(int $size, bool $unsigned = false, $zerofill = false)
+ * @method $this|AbstractColumn bigInteger(int $size, bool $unsigned = false, $zerofill = false)
+ * @method $this|AbstractColumn unsigned(bool $value)
+ * @method $this|AbstractColumn zerofill(bool $value)
  */
 class MySQLColumn extends AbstractColumn
 {
@@ -28,7 +38,7 @@ class MySQLColumn extends AbstractColumn
      */
     public const DATETIME_NOW = 'CURRENT_TIMESTAMP';
 
-    protected const ENGINE_INTEGER_TYPES = ['tinyint', 'smallint', 'mediumint', 'int', 'bigint'];
+    protected const INTEGER_TYPES = ['tinyint', 'smallint', 'mediumint', 'int', 'bigint'];
 
     protected array $mapping = [
         //Primary sequences
@@ -53,10 +63,10 @@ class MySQLColumn extends AbstractColumn
 
         //Integer types (size can always be changed with size method), longInteger has method alias
         //bigInteger
-        'integer'     => ['type' => 'int', 'size' => 11],
-        'tinyInteger' => ['type' => 'tinyint', 'size' => 4],
-        'smallInteger'=> ['type' => 'smallint', 'size' => 6],
-        'bigInteger'  => ['type' => 'bigint', 'size' => 20],
+        'integer'     => ['type' => 'int', 'size' => 11, 'unsigned' => false, 'zerofill' => false],
+        'tinyInteger' => ['type' => 'tinyint', 'size' => 4, 'unsigned' => false, 'zerofill' => false],
+        'smallInteger'=> ['type' => 'smallint', 'size' => 6, 'unsigned' => false, 'zerofill' => false],
+        'bigInteger'  => ['type' => 'bigint', 'size' => 20, 'unsigned' => false, 'zerofill' => false],
 
         //String with specified length (mapped via method)
         'string'      => ['type' => 'varchar', 'size' => 255],
@@ -91,7 +101,7 @@ class MySQLColumn extends AbstractColumn
 
     protected array $reverseMapping = [
         'primary'     => [['type' => 'int', 'autoIncrement' => true]],
-        'bigPrimary'  => ['serial', ['type' => 'bigint', 'autoIncrement' => true]],
+        'bigPrimary'  => ['serial', ['type' => 'bigint', 'size' => 20, 'autoIncrement' => true]],
         'enum'        => ['enum'],
         'boolean'     => ['bool', 'boolean', ['type' => 'tinyint', 'size' => 1]],
         'integer'     => ['int', 'integer', 'mediumint'],
@@ -122,7 +132,7 @@ class MySQLColumn extends AbstractColumn
         'mediumtext',
         'tinytext',
         'longtext',
-        'blog',
+        'blob',
         'tinyblob',
         'longblob',
     ];
@@ -130,13 +140,30 @@ class MySQLColumn extends AbstractColumn
     /**
      * Column is auto incremental.
      */
+    #[ColumnAttribute(self::INTEGER_TYPES)]
     protected bool $autoIncrement = false;
+
+    /**
+     * Unsigned integer type. Related to {@see INTEGER_TYPES} only.
+     */
+    #[ColumnAttribute(self::INTEGER_TYPES)]
+    protected bool $unsigned = false;
+
+    /**
+     * Zerofill option. Related to {@see INTEGER_TYPES} only.
+     */
+    #[ColumnAttribute(self::INTEGER_TYPES)]
+    protected bool $zerofill = false;
 
     /**
      * @psalm-return non-empty-string
      */
     public function sqlStatement(DriverInterface $driver): string
     {
+        if (\in_array($this->type, self::INTEGER_TYPES, true)) {
+            return $this->sqlStatementInteger($driver);
+        }
+
         $defaultValue = $this->defaultValue;
 
         if (\in_array($this->type, $this->forbiddenDefaults, true)) {
@@ -144,21 +171,14 @@ class MySQLColumn extends AbstractColumn
             $this->defaultValue = null;
         }
 
-        $statementParts = parent::sqlStatementParts($driver);
-
-        if (in_array($this->type, self::ENGINE_INTEGER_TYPES)) {
-            $attr = array_filter(array_intersect_key($this->attributes, ['unsigned' => false, 'zerofill' => false]));
-            if ($attr) {
-                array_splice($statementParts, 3, 0, array_keys($attr));
-            }
-        }
+        $statement = parent::sqlStatement($driver);
 
         $this->defaultValue = $defaultValue;
         if ($this->autoIncrement) {
-            $statementParts[] = 'AUTO_INCREMENT';
+            return "{$statement} AUTO_INCREMENT";
         }
 
-        return implode(' ', $statementParts);
+        return $statement;
     }
 
     /**
@@ -175,7 +195,7 @@ class MySQLColumn extends AbstractColumn
 
         if (
             !preg_match(
-                '/^(?P<type>[a-z]+)(?:\((?P<options>[^\)]+)\))?(?: (?P<attr>[a-z ]+))?/',
+                '/^(?P<type>[a-z]+)(?:\((?P<options>[^)]+)\))?(?: (?P<attr>[a-z ]+))?/',
                 $column->type,
                 $matches
             )
@@ -188,7 +208,7 @@ class MySQLColumn extends AbstractColumn
 
         $options = [];
         if (!empty($matches['options'])) {
-            $options = explode(',', $matches['options']);
+            $options = \explode(',', $matches['options']);
 
             if (count($options) > 1) {
                 $column->precision = (int)$options[0];
@@ -199,19 +219,19 @@ class MySQLColumn extends AbstractColumn
         }
 
         if (!empty($matches['attr'])) {
-            if (in_array($column->type, self::ENGINE_INTEGER_TYPES)) {
+            if (\in_array($column->type, self::INTEGER_TYPES, true)) {
                 $intAttr = array_map('trim', explode(' ', $matches['attr']));
-                if (in_array('unsigned', $intAttr)) {
-                    $column->attributes['unsigned'] = true;
+                if (\in_array('unsigned', $intAttr, true)) {
+                    $column->unsigned = true;
                 }
-                if (in_array('zerofill', $intAttr)) {
-                    $column->attributes['zerofill'] = true;
+                if (\in_array('zerofill', $intAttr, true)) {
+                    $column->zerofill = true;
                 }
                 unset($intAttr);
             }
         }
 
-        // since 8.0 database does not provide size for some of the columns
+        // since 8.0 database does not provide size for some columns
         if ($column->size === 0) {
             switch ($column->type) {
                 case 'int':
@@ -231,7 +251,7 @@ class MySQLColumn extends AbstractColumn
 
         //Fetching enum values
         if ($options !== [] && $column->getAbstractType() === 'enum') {
-            $column->enumValues = array_map(static fn ($value) => trim($value, $value[0]), $options);
+            $column->enumValues = \array_map(static fn ($value) => trim($value, $value[0]), $options);
 
             return $column;
         }
@@ -256,20 +276,17 @@ class MySQLColumn extends AbstractColumn
     public function compare(AbstractColumn $initial): bool
     {
         assert($initial instanceof self);
-        if (!parent::compare($initial)) {
-            return false;
-        }
+        return ! (!parent::compare($initial));
+    }
 
-        if (in_array($this->type, self::ENGINE_INTEGER_TYPES)) {
-            $attr = ['unsigned' => false, 'zerofill' => false];
-            foreach ($attr as $a => $def) {
-                if (($this->attributes[$a] ?? $def) !== ($initial->attributes[$a] ?? $def)) {
-                    return false;
-                }
-            }
-        }
+    public function isUnsigned(): bool
+    {
+        return $this->unsigned;
+    }
 
-        return true;
+    public function isZerofill(): bool
+    {
+        return $this->zerofill;
     }
 
     /**
@@ -288,5 +305,20 @@ class MySQLColumn extends AbstractColumn
         }
 
         return parent::formatDatetime($type, $value);
+    }
+
+    private function sqlStatementInteger(DriverInterface $driver): string
+    {
+        return \sprintf(
+            '%s %s(%s)%s%s%s%s%s',
+            $driver->identifier($this->name),
+            $this->type,
+            $this->size,
+            $this->unsigned ? ' UNSIGNED' : '',
+            $this->zerofill ? ' ZEROFILL' : '',
+            $this->nullable ? ' NULL' : ' NOT NULL',
+            $this->defaultValue !== null ? " DEFAULT {$this->quoteDefault($driver)}" : '',
+            $this->autoIncrement ? ' AUTO_INCREMENT' : ''
+        );
     }
 }