+++ src/Driver/SQLite/Schema/SQLiteColumn.php
@@ -28,6 +28,7 @@ class SQLiteColumn extends AbstractColumn
         'userType',
         'timezone',
         'size',
+        'attributes',
     ];
 
     protected array $mapping = [
@@ -158,7 +159,7 @@ class SQLiteColumn extends AbstractColumn
         $column = new self($table, $schema['name'], $timezone);
 
         $column->nullable = !$schema['notnull'];
-        $column->type = $schema['type'];
+        $column->type = \strtolower($schema['type']);
 
         if ((bool)$schema['pk'] && $column->type === 'integer') {
             $column->primaryKey = true;